working on prob 1 and 3

+++ Problem 3/stepCounter.py
@@ -1,6 +1,6 @@
 def feet_to_steps(user_feet):
    #write your code here
-   return(user_feet/2.5)
+   return(int(user_feet/2.5))
 
 if __name__ == '__main__':
     #take input feet steps here
@@ -8,4 +8,4 @@ if __name__ == '__main__':
     #store it into the function
     steps = feet_to_steps(user_feet)
     #print your steps here
-    print(int(steps))
+    print(steps)
